review changes - renamed HeaderDropdown
